+++ CHANGELOG.md
@@ -559,3 +559,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.178] - 2020-04-20
 ### Changed
 - Fixed wrong colors, font weights and css improvements.
+
+## [1.0.179] - 2020-04-20
+### Added
+- Display name for Label and Input components.
+
+### Changed
+- Label prop Types for children.